Refactored maybe-upgrade to remove construction
Set up a 'can-upgrade-now?' method similar to 'can-make-now?'
Had to create a can-upgrade-kw? in interop, because the can-upgrade?
accepted a unit type rather than a kw like can-make? does.
(I didn't want to change can-upgrade? incase it was set up that way
on purpose.)

Also changed building-built? to num-buildings-of-kw, as requested,
and fixed contracted-kw.